requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
numpy
pandas
tqdm
PyPDF2
glob2
